[DEV]: refactor
1. create common project
2. only allow view own photos
