+++ moai-java-ak47-common/src/main/java/com/github/wgx731/ak47/model/Photo.java
