+++ moai-java-ak47-ui/src/main/java/com/github/wgx731/ak47/vaadin/config/ServiceInitListener.java
