+++ moai-java-ak47-ui/src/main/resources/application.properties
@@ -1,0 +1,21 @@
+server.port=8800
+####### SPRING DATA PAGING ##############
+spring.data.web.pageable.default-page-size=20
+spring.data.web.pageable.max-page-size=100
+spring.data.web.pageable.one-indexed-parameters=false
+####### SPRING BOOT ADMIN ##############
+spring.boot.admin.client.url=http://localhost:8801
+management.health.ldap.enabled=false
+management.endpoints.web.exposure.include=*
+####### VAADIN ##############
+vaadin.compatibilityMode=true
+####### LDAP ##############
+ldap.enabled=true
+ldap.urls=ldap://ldap.forumsys.com:389/
+ldap.base.dn=dc=example,dc=com
+ldap.username=cn=read-only-admin,dc=example,dc=com
+ldap.password=password
+ldap.user.dn.pattern=uid={0}
+####### LOCAL ##############
+local.user.password={sha256}97cde38028ad898ebc02e690819fa220e88c62e0699403e94fff291cfffaf8410849f27605abcbc0
+local.admin.password={sha256}97cde38028ad898ebc02e690819fa220e88c62e0699403e94fff291cfffaf8410849f27605abcbc0
