+++ moai-java-ak47-ui/src/test/java/com/github/wgx731/ak47/vaadin/view/PhotoEditorTest.java
