--- moai-java-photo-web/src/main/java/com/github/wgx731/ak47/repository/PhotoRepository.java
@@ -1,7 +1,0 @@
-package com.github.wgx731.ak47.repository;
-
-import com.github.wgx731.ak47.model.Photo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PhotoRepository extends JpaRepository<Photo, Long> {
-}
