--- moai-java-photo-web/src/main/resources/application.properties
@@ -1,18 +1,0 @@
-server.port=8800
-####### SPRING DATA PAGING ##############
-spring.data.web.pageable.default-page-size=20
-spring.data.web.pageable.max-page-size=100
-spring.data.web.pageable.one-indexed-parameters=false
-####### SPRING BOOT ADMIN ##############
-spring.boot.admin.client.url=http://localhost:8801
-management.health.ldap.enabled=false
-management.endpoints.web.exposure.include=*
-####### VAADIN ##############
-vaadin.compatibilityMode=true
-####### LDAP ##############
-ldap.enabled=true
-ldap.urls=ldap://ldap.forumsys.com:389/
-ldap.base.dn=dc=example,dc=com
-ldap.username=cn=read-only-admin,dc=example,dc=com
-ldap.password=password
-ldap.user.dn.pattern=uid={0}
