+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'moai-world'
 
 include 'moai-java-threading'
-include 'moai-java-photo-web'
 include 'moai-java-spring-admin'
+include 'moai-java-ak47-ui'
+include 'moai-java-ak47-common'
